Allow a dragStepSize function for a Series to dynamically evaluate drag step sizes of x/y. The dragStepSize function gets two parameters: * XorY, specifying with 'X' or 'Y' if what axis it relates to * point, being the point that is dragged
Allowed a dragStepSizeX and dragStepSizeY option for specifying static number drag step sizes.

Added dragStepAllowMinMax and dragStepRelative that modifies the behavior of the previously mentioned options.
dragStepAllowMinMax will allow/not allow going to min/max when dragging, if it does not match a step. Defaults to false.
dragStepRelative will make the steps relative to the original value of the point being dragged.